Rename models to metadata
